# bkn-trace/agent-observability/docs/swagger/swagger.yaml
swagger: '2.0'
info:
  description: APIs for querying agent traces from OpenSearch.
  title: Agent Observability API
  contact: {}
  version: '1.0'
basePath: /api/agent-observability/v1
paths:
...
  /requests:
    get:
      description: Returns stable request summaries generated from authorized evidence and artifacts.
      produces:
      - application/json
      tags:
      - requests
      summary: List observable business requests
      parameters:
      - type: integer
        description: Page size, 1..200
        name: limit
        in: query
      - type: string
        description: Opaque pagination cursor
        name: cursor
        in: query
      - type: string
        description: One user interaction ID
        name: interaction_id
        in: query
      - type: string
        description: Caller-owned conversation ID
        name: conversation_id
        in: query
      - type: string
        description: Started at or after this RFC3339 timestamp
        name: from
        in: query
      - type: string
        description: Started at or before this RFC3339 timestamp
        name: to
        in: query
      - type: string
        description: Execution status
        name: status
        in: query
      - type: string
        description: Agent or application
        name: agent_or_app
        in: query
      - type: string
        description: Business domain
        name: business_domain
        in: query
      - type: string
        description: Knowledge network
        name: knowledge_network
        in: query
      - type: string
        description: Evidence completeness
        name: evidence_completeness
        in: query
      - type: string
        description: Question, result, ID, business ref, or error keyword
        name: keyword
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/evidencevo.RequestSummaryPage'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
  /requests/{request_id}:
    get:
      description: Returns an authorized request summary and its business-content availability.
      produces:
      - application/json
      tags:
      - requests
      summary: Get one observable business request
      parameters:
      - type: string
        description: BKN request ID
        name: request_id
        in: path
        required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/evidencevo.RequestSummary'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
  /requests/{request_id}/traces:
    get:
      description: Supports one business request to multiple distributed traces and returns request_id on every trace.
      produces:
      - application/json
      tags:
      - requests
      summary: List technical traces belonging to a business request
      parameters:
      - type: string
        description: BKN request ID
        name: request_id
        in: path
        required: true
      - type: integer
        description: Page size, 1..200
        name: limit
        in: query
      - type: string
        description: Opaque pagination cursor
        name: cursor
        in: query
      - type: string
        description: One user interaction ID
        name: interaction_id
        in: query
      - type: string
        description: Caller-owned conversation ID
        name: conversation_id
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/evidencevo.TraceSummaryPage'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
  /trace-executions:
    get:
      description: Returns stable trace summaries with reverse request links, generated from authorized trace evidence.
      produces:
      - application/json
      tags:
      - traces
      summary: List technical trace executions
      parameters:
      - type: integer
        description: Page size, 1..200
        name: limit
        in: query
      - type: string
        description: Opaque pagination cursor
        name: cursor
        in: query
      - type: string
        description: One user interaction ID
        name: interaction_id
        in: query
      - type: string
        description: Caller-owned conversation ID
        name: conversation_id
        in: query
      - type: string
        description: Exact trace ID; bypasses list projection scan
        name: trace_id
        in: query
      - type: string
        description: Started at or after this RFC3339 timestamp
        name: from
        in: query
      - type: string
        description: Started at or before this RFC3339 timestamp
        name: to
        in: query
      - type: string
        description: Execution status
        name: status
        in: query
      - type: string
        description: Agent or application
        name: agent_or_app
        in: query
      - type: string
        description: Business domain
        name: business_domain
        in: query
      - type: string
        description: Trace, request, operation, or error keyword
        name: keyword
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/evidencevo.TraceSummaryPage'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
...
  /interactions/{interaction_id}:
    get:
      description: Aggregates every authorized request and trace in one caller-owned interaction.
      produces:
      - application/json
      tags:
      - interactions
      summary: Get one observable business interaction
      parameters:
      - type: string
        description: BKN interaction ID
        name: interaction_id
        in: path
        required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/evidencevo.InteractionSummary'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rdto.ErrorResponse'
definitions:
  evidencevo.ActionSummary:
    type: object
    properties:
      approved:
        type: integer
      completed:
        type: integer
      executed:
        type: integer
      last_status:
        type: string
      recommended:
        type: integer
...
  evidencevo.RequestSummary:
    type: object
    properties:
      action_summary:
        $ref: '#/definitions/evidencevo.ActionSummary'
      agent_or_app:
        type: string
      business_domain:
        type: string
      business_refs:
        type: array
        items:
          type: string
      completed_at:
        type: string
      conversation_id:
        type: string
      duration_ms:
        type: integer
      error_summary:
        type: string
      interaction_id:
        type: string
      evidence_completeness:
        type: string
      initiator:
        type: string
      knowledge_networks:
        type: array
        items:
          type: string
      partial_reasons:
        type: array
        items:
          type: string
      question_preview:
        type: string
      request_id:
        type: string
      result_preview:
        type: string
      started_at:
        type: string
      status:
        type: string
      trace_count:
        type: integer
  evidencevo.RequestSummaryPage:
    type: object
    properties:
      entries:
        type: array
        items:
          $ref: '#/definitions/evidencevo.RequestSummary'
      next_cursor:
        type: string
      partial:
        type: boolean
      partial_reasons:
        type: array
        items:
          type: string
      total:
        type: integer
      truncated:
        type: boolean
...
  evidencevo.TraceSummary:
    type: object
    properties:
      agent_or_app:
        type: string
      business_domain:
        type: string
      completed_at:
        type: string
      conversation_id:
        type: string
      duration_ms:
        type: integer
      error_summary:
        type: string
      interaction_id:
        type: string
      request_id:
        type: string
      root_operation:
        type: string
      span_count:
        type: integer
      started_at:
        type: string
      status:
        type: string
      trace_id:
        type: string
  evidencevo.TraceSummaryPage:
    type: object
    properties:
      entries:
        type: array
        items:
          $ref: '#/definitions/evidencevo.TraceSummary'
      next_cursor:
        type: string
      partial:
        type: boolean
      partial_reasons:
        type: array
        items:
          type: string
      total:
        type: integer
      truncated:
        type: boolean
...
  rdto.ErrorResponse:
    type: object
    properties:
      code:
        type: string
      details: {}
      error_code:
        type: string
      message:
        type: string
      trace_id:
        type: string
  evidencevo.InteractionSummary:
    type: object
    properties:
      completed_at:
        type: string
      conversation_id:
        type: string
      duration_ms:
        type: integer
      interaction_id:
        type: string
      requests:
        type: array
        items:
          $ref: '#/definitions/evidencevo.RequestSummary'
      started_at:
        type: string
      status:
        type: string
      traces:
        type: array
        items:
          $ref: '#/definitions/evidencevo.TraceSummary'